JSParser: replace AndroidJSCore with Rhino
AndroidJSCore is quite a pig because of it's jni libs (adds 40+MB to app)
and I haven't figured out a way to terminate an executing script.

Webview is present already on Android but I haven't been able to get it to
work in a non-UI thread (likely because its used in the UI thread by blockly
and I don't think you can have multiple webviews in differring threads).
Regardless, its a pig performance-wise and a pain to deal with if needing to
run in non UI thread just for javascript parsing.

Rhino seems like a great choice:
 - it's jar is 1.5MB
 - you can easily hook in a debugger to identify what line is being executed
   and to easily terminate a script
 - it can run in any thread
 - it doesn't need any UI

Signed-off-by: Tim Harvey <harvey.tim@gmail.com>